weather_app/requirements.txt
Django
pytest
pytest-pythonpath
pytest-django
djangorestframework
djangorestframework-simplejwt
Pillow
PyJWT
requests
asgiref
adrf
python-dotenv
psycopg2
httpx
asyncio
python-decouple==3.5
